Fix left pointer of 28 bit records.
The left nibble of the middle byte in the 28 bit format was not
correctly shifted prior to combining it with the other 3 bytes of
the left pointer. Namely in the following 28 bit node structure
where XXXX denotes the most-significant bits of the left pointer

LLLLLLLL LLLLLLLL LLLLLLLL XXXXRRRR RRRRRRRR RRRRRRRR RRRRRRRR

the left pointer would end up incorrectly as

XXXX0000 LLLLLLLL LLLLLLLL LLLLLLLL

instead of

0000XXXX LLLLLLLL LLLLLLLL LLLLLLLL

due to a missing shift of 4 bits.